requirements.txt
decord
hydra-core
numpy
pyyaml
safetensors
torch
torchaudio
tqdm
